Refactor od function
